Switched to webkit listeners for faster recognition

+++ jeeves.py
@@ -1,22 +1,28 @@
 from __future__ import print_function
+from twisted.internet import reactor
 from segregator.segregator import Segregator
-from google_speech_api.recognizer import recognize
+from webkit_speech_api.recognizer import recognize
 # from julius_connector.julius_connector import connect_to_julius
+# from google_speech_api.recognizer import recognize
 
 import argparse
 import subprocess
 import signal
 import sys
 import os
-from utils.user_input import user_input
 
 
 def signal_handler(signal, frame):
-    with open(".CHAT_SERVER_PID") as pid:
-        subprocess.call(['kill', pid.readlines()[0].strip()])
+    if os.path.exists(".CHAT_SERVER_PID"):
+        with open(".CHAT_SERVER_PID") as pid:
+            subprocess.call(['kill', pid.readlines()[0].strip()])
+            os.remove(".CHAT_SERVER_PID")
 
-    os.remove(".CHAT_SERVER_PID")
-    exit(0)
+    reactor.stop()
+    try:
+        exit(0)
+    except SystemExit:
+        print("\n<<< Jeeves Terminated >>>")
 
 
 signal.signal(signal.SIGINT, signal_handler)
@@ -33,13 +39,14 @@ if not os.path.exists(".CHAT_SERVER_PID"):
     with open(".CHAT_SERVER_PID", "w") as pid:
         pid.write(str(process.pid))
 
-src = "mic" if (args.input and args.input == "mic") else "stdin"
+src = "stdin"
+if args.input and args.input == "mic":
+    src = "mic"
 
-while True:
-    if src == "stdin":
+if src == "stdin":
+    while True:
         sentence = raw_input("Type something: ")
-    else:
-        sentence = user_input("Say something: ")
-
-    segregator = Segregator(sentence)
-    segregator.segregate_and_react()
+        segregator = Segregator(sentence)
+        segregator.segregate_and_react()
+else:
+    recognize()
